Add support for Graphite/Statsite
This changeset adds support for three different metrics collection
tools:

- Collectd

Collectd is a lightweight system metrics (disk, memory, CPU) collection
agent.

- Graphite

Graphite is made up of several things, but in general it is a
time-series data store with dynamic graph generation capabilities.

- Statsite

Statsite is an implementation of StatsD in C. It aggregates incoming
statistics and flushes them to a backend time-series data store
(Graphite in this case).

All of these things are being added to the environment in order to
accomplish something like this:

http://codeascraft.com/2011/02/15/measure-anything-measure-everything/

In order to view the Graphite Web UI, navigate to
`http://localhost:8082` on the VM host. Also, standby for @jwalgran to
determine what client will be used to feed Statsite/Graphite from
Python.