Add test provisioning path
This changeset introduces an alternate provisioning path for testing
environments. The majority of the differences in a test provisioning
path are associated with the omission of monitoring (Collectd, RELP,
Logstash). There is a new `nyc-trees.monitoring` role that attempts to
encapsulate monitoring tasks associated with a metrics/logs producer.

The environmental variable `VAGRANT_ENV` is used to determine which
provisioning path is taken. Inside of Jenkins, `VAGRANT_ENV` will be set
to `TEST`, which will adjust the Ansible inventory used.