Use .galaxy_install_info to trigger role installation
The `ansible-galaxy` command is nice enough to leave behind a bit of
metadata associated with role installation
(`role_name/meta/.galaxy_install_info`). The `Vagrantfile` now attempts
to determine the version of an installed role based off of the metadata
in that file. If the versions don't match, an install of all role
dependencies occurs.

I'm not in love with this solution, but I think that it is better than
force reinstalling all role dependencies every time (~10 seconds on
every provision).