Move achievements into the user app
Achievements seem to fit better in the user app.
Achievements are shown on the user profile.
The achievements page will only be show the logged in users achievements

What was left in the app related to quizess so it is now the quiz app.

The AchievementDefinition model was also dropped in favor of a static list