Upgrade NPM
This changeset upgrades the version of NPM from 1.4.28 to 2.1.14. Joyent
bundles Node.js with the current version of NPM at time of release. In
our case, 0.10.32.

Locally, this change appears to have increased the probability of build
success. I'm going to do more testing, but putting this up here so that
others can confirm.