Add geocoder interface to map module
The geocode interface is split into three layers:

1. A generic ``ajaxLatest`` module that allows for interruptible
   requests.
2. A wrapper around the geocode AJAX requests that converts responses to
   events.
3  A controller that pulls markup from template script tags and manages
   the display of UI elements in response to user events and geocode events.

To keep the ``map`` module lean, I moved a lot of code initially written
in that module out to ``searchController``, but I needed access to map
zooming helpers that were initially added to ``map``. My solution was to
move the helpers out to a new ``mapUtils`` module.