Add partial email backend for AWS IAM Credentials
This is the culmination of the work @hectcastro and I did to get an
email backend working with AWS IAM role crdentials fetched internally
from the metadata server. It turns out that this is not viable because
one of the few cases not covered by these dynamic credentials is sending
email. However, after being thoroughly exercised, this code is polished
enough to support the feature if AWS enables it, or can be adapted to a
similar, working solution.