Add reservable layer to the reservation page
To guide a person in selecting blocks to add to their reservation, we
need a simple available/not available layer.

I am adding this to the initial view of the reservation page for
bootstrapping purposes only. When the "add" mode is implemented, it will
show this layer only in that mode.