Add bootstrapping of default training pages
fixes #310 on github indirectly.

This commit adds the scaffolding for the required training pages as well
as a management command to create them if they do not exist.

Issue #310 on github is now fixed because we don't mount flatpages to
the same endpoint as actual pages, so we can revert to a simpler URL
pattern for resolving flatpages. This removes the broken path problem that
cropped up in 310.