Add SES/boto based email backend for prod
Using boto on a properly provisioned EC2 keeps the application out of
the business of managing email credentials.

All boto interaction is isolated in the ``BotoMailer`` class, which has
no dependencies on Django and can be tested in isolation.

This backend is configured for production only because it requires
EC2 metadata services that aren't available on vagrant VMs.