Add HTTP caching support for tiles and static assets
This changeset adds HTTP caching support to production deployments by:

- Enabling the Cache-Control, Expires, and ETag headers for static assets
- Enabling the Cache-Control and Expires headers for tiles

It is worth noting that when the tiler is fronted by CloudFront, CloudFront will choose the greater of Cache-Control vs. 24 hours, but reserves the right to purge infrequently accessed assets from edge locations to make way to more frequently accessed ones.

Attempts to resolve #302 and #303.