Add ability to select blockfaces for reservation.
This is only the client-side portion of reserving blockfaces.
More work will need to be done in a future PR to post the blockface ids
to the server in order to reserve them.

Selected blockfaces will have a vector line laid overtop of them to
indicate that they are currently selected.

This was abstracted into selectableBlockfaceLayer.js for reuseability with
other pages where we will need to show the currently selected blockface(s)