Copy event URL to clipboard
This is a low-tech solution requiring the user to type Ctrl+C (or Cmd+C).
There are solutions relying on flash and 20K of JS that we could integrate.
They don't work on iOS. Doesn't seem worth it to me.
My solution's UI isn't pretty, but these users are admins.

Fixes #422