Add DNS support for monitoring server
This changeset gives the monitoring server a subdomain off of the
default Route 53 hosted zone.

For example, if the default hosted zone is `treescount.foo.com`,
these CloudFormation changes will create an A record of
`monitoring.treescount.foo.com` under the `treescount.foo.com` hosted
zone that points to the public IP address of the monitoring server.