Show reservations section on profile/dashboard
The markup matches the style suggested by the wireframes, but does not
include an address, because the blockface data being generated for the
census will not include an address.

I humanized the dates using ``naturalday`` to get expiration dates to
display as  'today' and 'tomorrow' when appropriate.

The reservation rows link to the reservation page with a blockface id
anchor so that the client side map code can zoom to and select the
blockface when the page loads.