Switch to using explicit markup instead of forms for registration pages.
Since the Django Forms for these pages come from contrib.auth or
django registration, it is not easy for us to make changes to the
generated markup.

Django forms are still used on the backend for validation, and we continue
to use Django forms in the markup for showing errors.