Fix layout issue when error message is displayed
It looks like this was somewhat fixed by recent design changes. I moved
the validation message outside of the input div so that the error
message does not increase the length of the input field.

Also, I swapped field labels/errors because IMO it makes it much clearer to
see which form fields the errors are talking about.

There is remaining work to style validation messages.

Fixes #492