Admin page for managing group territory
Show these categories of blockface:
* Available
* Reserved by this group
* Reserved by others
* Mapped by this group
* Mapped by others

These are essentially the same categories as shown on the "Reservations" page.

Initially we thought all of the following should be shown as "unavailable":
* Reserved by others
* Mapped by this group
* Mapped by others

I wanted to
* distinguish "Mapped by this group" to give context when working on a group,
  for example if you want to grow territory contiguously.
* distinguish "Reserved by others" from "Mapped by others" to allow seeing
  what yet needs to be mapped in a reapportioning situation.

With good coloring I think these will be useful and not distracting.
And it's easy to go back to 3 categories if we prefer.

Also:
* Fix omission of `BlockfaceReservation` in `get_context_for_territory_survey_layer`
* In `mapModule.addTileLayer` and `addGridLayer`, replace the unused `domId` argument with
  a `queryString` argument so the client can add query arguments to tiler URLs