Add a popup for exiting a survey without mapping
This change set allows a mapper to "bail out" of a survey, and forces
them to leave an explanation as to why they are quitting. I have opted
to not post any tree data because it could be invalid, and it could be
frustrating to a person who is trying to stop mapping if we force them
to make data corrections first.

What happens after the quit has been POSTed is left as a TODO because
that part of the workflow is in progress.