Add teammate selection to survey workflow
The teammate list is implemented as a select tag because we intend to
use a UI library that converts select tags to autocomplete lists.

Respects privacy settings and shows full names only when allowed.

Adding another set of fields to the existing blockface selection
workflow made the hide and show logic a little too messy so I added a
helper that builds functions to show and hide mutually exclusive field
groups.