Remove reservations when assigning expert blocks
If a block is being reassigned to the Parks group for expert handling,
existing reservations should be deleted. It would be more correct to
check who reserved the block and preserve the reservation if they are,
indeed, and expert, but this is an edge case and not worth the
development time to cover.