Add support for setting Django site domain
This change adds support for setting the `name` and `domain` of the
default site in `django_site`. Support has been added via `setupdb.sh`,
which depends on the existence of a Django management command called
`set_django_site_domain`.

If no name or domain is passed to `setupdb.sh` or the Django management
command, then it defaults to `treescount.nycgovparks.org`.