Add Celery support to application server
This changeset adds Celery support to the application servers by wiring
up a Redis broker and launching a Celery worker on each application
server. The workers aren't tied to the application server they colocate
with; tasks still flow through the a broker, and that broker attempts
to evenly distribute work.

In addition, Celery signals are configured to emit metrics about job
count and duration to Graphite via StatsD/Statsite.

Attempts to resolve #851.