Send email when trusted mapper is approved
Individual mappers request access to map in a group's territory. If a
group admin approves that request, we notify the user and send a link to
the reservation page.

Because we are going to have several message types, some on demand, some
scheduled, I opted to put as keep much of the email logic as possible,
along with the message templates, in a new "mail" app rather
than start scattering them around the application.