Add support for Route 53 failover to static S3 website
If all of the nodes behind the application load balancer happen to fail,
Route 53 will begin passing requests through to the failover DNS
endpoint. In this case, the failover endpoint is a static S3 website,
fronted by CloudFront, with SSL certificates that match the application
load balancers.

The CloudFront dependency is necessary because we're steering users
through HTTPS to interact with the application. The S3 website endpoints
are unable to handle HTTPS, which could lead to users seeing connection
errors when they should really be seeing the failover website.
CloudFront supports HTTPS connections with our SSL certificates via SNI.