Add management command to dump tables to CSVs
The command takes an optional argument, which will be the name of the
directory, under the default Django storage backend, into which the CSV
files will be written.

I wrote the export as a plain function call, but ensured that all the
arguments are serializable, so that if the need for async processing
arises we can easy switch to celery.