Expand functionality of event list filters
Add support for dynamic empty messages, including embedded urls. Also
provide the whole request context to the filter functions.

Necessary to add more advanced filters and empty messages.

Add support for conditional visibility of filters even if they are part
of the chosen filterset. We turned out not to use this, but it will be
useful if we expand on event lists, which has happened before.