Attach event PDF to reservation confirmation email
Since the meeting point for an event may not be easily expressed as an
address, it helps to send the registrant a map centered on the event
location.

I made the ``send_to`` helper expect and handle a list of attachments
because reservation confirmation emails will need to include both a
static data collection template and the dynamic reservation map PDF.

Is is possible that the PDF for a new or updated event is has not been
created by the time a person RSVPs for the event, so we need to chain
the notification task after ``wait_for_default_storage_file``.