Expose tiler metatile count as config option
The metatile setting has a dramatic effect on both real world and
perceived tile rendering speed. A larger value may be appropriate for a
single tile rendering process but a smaller one be appropriate for
multiple tile rendering processes running in a load balanced
cluster. Modifying the server to read the metatile count from the
environment allows for experimentation and performance tuning without
code modification or redeployment.