Don't fetch selected blockface if onAdd is falsey
The ``onAdd`` override gives client code a chance to choose whether or
not the blockface will be highlighted on click. The commit makes it so
that the blockface is fetched only if ``onAdd`` returns truthy.

The geom is no longer available to be passed to ``onAdd``, but it was
not used by any of the implementers, so it could be safely removed from
the signature.

Also includes a documentation update to reflect revised behavior.