Don't ignore event related HTTP errors on survey pages
The `_validate_event_and_group` function returns an HTTP error if the
user is not checked into the current event. Consumers of this helper function
erroneously assume that a valid event is returned. The result is that some
survey endpoints can be triggered by users even if they have not been
checked into the event.