Bump pip role version
This change fixes a scenario during AMI creation where Packer installs
the `python-pip` and `python-dev` packages prior to running Ansible
because Python is an Ansible dependency. The version of the `python-pip`
package installed when no version number is specified (as is the case in
the Packer template) conflicts with the version installed by this role
because it prompts the user for dependent package downgrades.