Load admin page OpenLayers from https CDN
The Django admin has a built in slippy map for editing geometries, built
on top of OpenLayers. By default, the admin template loads OpenLayers
from http://openlayers.org/api/2.13/OpenLayers.js, which fails with an
"insecure" error when the site is running under https.

This commit overrides that insecure URL with a secure one from cdnjs.