Add Nginx caching layer for Windshaft
This changeset adds a file system based proxy cache (provided by Nginx)
to Windshaft. Windshaft is responsible for providing `Cache-Control`
headers to Nginx (and ultimately CloudFront) so that it can determine
whether to cache the response or not.

In addition, EC2 ephemeral storage is used (when available) as the file
system destination for cached items. `/mnt` is commonly the path where
ephemeral storage is mounted.