Enable Windshaft profiler in development; add JMeter test plans
These changes add Ansible variables for enabling the Windshaft profiler.
By default, the profiler will be enabled in development and off in all
other environments.

What does that mean? That means that Windshaft will emit more detailed
metrics to the configured StatsD endpoint. In our case, we run Statsite
(a C port of StatsD). Statsite accepts these metrics and flushes them to
Graphite, which has a UI available on the `services` virtual machine on
port `8080`.

In addition to the Windshaft profiler hooks, a new top level directory
has been added with Apache JMeter test plans. Each test plan includes a
series of requests that can be replayed with JMeter to simulate load
against the development environment. See the included `README` for more
details on how to maneuver JMeter.