Add activation email resend workflow
If a person has registered, but not activated their account, and they
attempt to log in, we want to tell them why we can't yet log them in,
and provide them a way to request another activation email.

This commit adds a workflow, similar to "forgot username," for resending
activation emails. There are a few key differences.

  * If the account is already active, there is no token, so we cannot
    send an email and, instead, prompt the person to log in.

  * Adding an anchor tag to the inactive account validation message
    requires overriding the authentication form, so that the HTML
    content can be marked as "safe."