Add group.territory_updated_at field and use for territory cache buster
Ensures that the cache buster is updated when blockfaces are unassigned
from a group's territory without any other changes taking place.

To test, remove a blockface from a group's territory and save without
making any other changes.  The color of the removed blockface should
change to indicate that it is now unavailable.

Fixes #791