Update gulp-ruby-sass to 1.0.5 to fix sourcemap and minification support.
Sourcemaps were broken in the version of gulp-ruby-sass we were using
because it wanted us to serve the original files and would not allow us
to choose the path to serve them from.

Minification was not implemented in favor of using a "compressed"
outputStyle because we did not have sourcemap support and I wanted to
enable debugging production issues easily.

Minification doesn't get us much (33Kb uncompressed vs. 30Kb compressed),
but it was trivial to add.

Fixes #142