Patch the select2 library to fix a bug on some Android browsers
There is a known bug with the 3.x version of Select2 that causes the
dropdown menu to sometimes immediately close after opening:
https://github.com/select2/select2/issues/2300
https://github.com/select2/select2/issues/2061

This only seems to occur on firefox for Android and the default Android
browser.

It seems to have been fixed in the 4.0 version of Select2, which was
released 2 days ago, but there are a number of breaking changes, so I'm
wary of upgrading to it.

Instead, I'm bringing in a fix I found in a fork of of the library, which
seems to fix the issue without causing any other effects.

Connects to #885