Create region fixtures, load on setup
Showing mapping progress aggregated by borough and neighborhood requires
that those polygons be available in the database. This commit adds
fixtures to load the source data and a management command to assign a
borough and neighborhood to each blockface.

While documenting the process I consolidated the existing data
processing documentation with a new readme in the survey fixtures
directory, since the data processing is similar.