Add indexes to 'updated_at' for Survey and BlockfaceReservation
These tables are queried to find the most recent update time when
computing tiler cache buster values, when loading that map page.

Survey is queried for the most recent 'updated_at' for every progress page
layer, and BlockfaceReservation.updated_at is queried for every other
map layer.

I am only adding indexes for these 2 table's 'updated_at' columns because
I expect that the number of rows in the other tables we use to calculate
cache busters will remain small, or already have indexes.

I attempted to measure the performance impact of these changes, but the
Postgresql query plan did not use the new indexes.  I suspect this is
because the amount of data in my development DB is rather small.

That said, I expect we will have a large number of rows in these tables
by the end of the census, perhaps numbering in the hundreds of thousands.

Connects to #275