Treecorder: sticky tree headers
In a scrolling list of sections, ensure that header of first visible section
is always visible. Works with add/remove section and collapse/expand section.

Adds either "fixed" or "absolute" class to a header when its section is
the first visible.

CSS class "fixed" uses position:fixed so the header won't scroll.

Class "absolute" is used when the header is at the bottom of its section.
CSS uses position:absolute so the header will scroll off the top.

Each header is wrapped by a <div>, used to preserve its height when
its positioning changes.