Add AutoScalingPolicy to application and tile servers
There are several high traffic periods in the day when mapping events
occur. Typically, these happen in bursts and then taper off after an
hour or two. In order to protect ourselves during these surges of
traffic, scale up and scale down policies have been added to the
application and tile servers.

Looking at the ELB graphs for overall request count, a threshold of 60
with a period of 60 seconds and 2 consecutive evaluation periods is a
good combination to trigger a scale up event. Before the following
minute is over, server capacity doubles.

On the other hand, a threshold of 75 with a period of 60 seconds and 10
consecutive evaluation periods is a safe combination to trigger a scale
down event. This is more conservative and protects against lulls
between evaluation periods where activity drops, but then picks back up
in the next minute or two.

Regardless, the desired and minimum capacity for each AutoScalingGroup
remains at 2.

Fix