Turn off zoom fixes on the progress map when reloading any layer
We need to put events on zoomend for the all progress view, but it
interferes with the selection on the user and group views.

We had been explicitly turning the zoomend handler off, but it was only
affecting user progress, not group progress.  Moving the .off(..) calls
should make it turn off for both.

Connects to #1697