Fix field validation for the "Distance to End" field
The validation for "Distance to End" was broken by the recent iOS fixes
that scroll the page on validation failures.  The JS was trying to scroll
to the fieldset containing the input that failed validation, but the
Disatnce to End field is not inside a fieldset.

To fix this and prevent future regressions, I have changed it so that we
fall back to scrolling to the input itself if we cannot find a fieldset.

Connects to #1701