Fix issue with duplicate IDs on treecorder.
Commit 8b4ee9db added IDs to several elements for styling purposes, but
because tree_form.html is used for client-side templating,
this lead to duplicate IDs once a tree is added to the page, which is
problematic for a number of browsers.

The fix is to template the id with the tree number, which we are already
doing for several elements in the same file.