Switch from using builtin browser validation messages to custom messages.
We need to show custom error messages on iOS and older Android browsers,
as they will not do so.  Since we are doing so for some browsers, I have
opted to switch to custom error messages on all browsers.

We still rely on html5 for the actual validation checking, which does work
on webkit.

Connects to #1664