Prevent unassigning group territory for surveyed block edges.
There was a bug in the view for administering census zones, causing
surveyed block edges to be unassigned when their group's territory was
updated.

This fixes the original issue, and contains a data migration to reassign
those blocks back to their group.

The data migration finds all blocks with no reservations and no territory,
and if they are inside a group polygon, adds them to that group's
territory.  Blocks which are not in any group polygons are assumed to
belong to the 'NYC Parks' group.

Connects to #1717