Update CloudFormation templates so that resources are up-to-date
In the process of creating a new stack from an existing template, the
following resources and their parameters became stale and required
updates:

- CloudFront Distribution/MinimumProtocolVersion

If you specify the `IamCertificateId` property and specify SNI only for
the `SslSupportMethod` property, you must use `TLSv1` for the minimum
protocol version. If you don't specify a value, AWS CloudFormation
specifies SSLv3.

- AutoScalingGroup/NotificationConfigurations

`NotificationConfiguration` is now `NotificationConfigurations` (plural)
and now accepts an array of `NotificationConfigurations`.