Use stamen toner labels on top of a label-less basemap for most maps.
This is intended to prevent the blockface tile layer from obscuring the
street names, to help in identifying the correct blockface while mapping.

I kept the progress map as it was, because I was running into issues
with labels interfering with the labels on the aggregate neighborhood
and borough layers.

I also excluded the event list map and the survey detail map, because
they do not have a blockface layer, so there is no need for a separate
label layer.

Thanks to @jwalgran for the initial implementation!

Connects to #1759