Fix label positions on the progress map for concave polygons.
Some of the neighborhood polygons on the progress map have centroids that
lie outside the polygon, which leads to their percentage label being
placed inside a different neighborhood (which will then have two labels).

By adding `text-placement: interior` to the CartoCSS, we tell Mapnik to
prefer placing labels on the interior of polygons if the centroid is not
inside the poygon.