Prevent duplicates of email addresses for different users usernames
This disallows signing up with a username which is already in use as an
email address, and vice versa.

This is to prevent a scenario where it would be ambiguous as to which
user we are trying to authenticate, since we are now allowing login by
either username or email.