Fix event map PDF generation triggered by logged out users
Shelling out to PhantomJS happens within a function that takes session
ID as an argument. That session ID is then passed to PhantomJS via
command line arguments.

For reservation PDFs, the session ID is needed to tailor the map for the
user who reserved. For event PDFs, no session ID is needed, but it still
goes through the same function as reservation PDFs. In cases where the
user who triggered an event PDF regeneration is logged out, the session
ID is `None`, which `execv` (inside `subprocess.check_output`) doesn't
like. It expects all arguments to be Python strings.