Count quit surveys as completed for aggregate counts.
Surveys quit for a specific reason (intended for things which prevent
mapping the block edge, like construction) are now considered as
completed for the purposes of aggregate counts such as percent completed
on the home page and the block and neighborhood percentages on the
progress map.

For the more detailed view of the all progress layer, and for the user and
group progress layers, we show quit blocks as gray, and the legend shows
that they could not be mapped.

On the event detail map, I chose to quit blocks the same color as mapped
blocks, since the more pertinent information there is what is available.

Connects to #1800