Add support for limiting access to the site after the census ends.
Uses the existing soft launch infrastructure to support full access to
the site when the census is suspended to a limited group of people, and
allow some access for other users.

Users who do not have full access can get to the progress map, group and
event lists, group and event detail pages, the FAQ and the main training
page (but not the training subpages).

Access to editing groups and events is blocked off unless the user has
full site access, as is access to following groups.

Connects to #1814