Bump django-statsd-mozilla to 0.3.15
Version 0.3.14 of this dependency did not pin its dependency for
`statsd`:

https://github.com/andymckay/django-statsd/commit/7f39cbcbaa12145db04eb4a0625150a24f5678b4#diff-b4ef698db8ca845e5845c4618278f29aL3

This allowed `statsd==3.2` to be installed, which introduces a new
constructor to determine IPv6 support:

https://github.com/jsocol/pystatsd/commit/cc56f1cfb29ff480e3db2cf39447b8232231c357#diff-398adcc6ad57dace01c71ba765896ebbR134

The placement of that argument is now where metric `prefix` used to be:

https://github.com/andymckay/django-statsd/blob/2e092e2269b097d0a5f4728db0198bbf09b6f38d/django_statsd/clients/__init__.py#L29

We're using the prefix `django`, which is truthy and triggers IPv6
support. Enabling IPv6 support makes the connection to
`monitoring.service.nyc-trees.internal` fail on EC2.