Add remaining blockedge fixture
These are the steps I used to generate a fixture from the
Remaining_Segments-08-11-15 shapefile received from Parks.

* Projected the shapefile to 4326 using ArcMap

* Loaded the shapefile into the database using shp2pgsql

    $ shp2pgsql -s 4326 -g geom \
    Remaining_Segments-08-11-15_4326.shp > \
    Remaining_Segments-08-11-15_4326.sql
    $ cat Remaining_Segments-08-11-15_4326.sql \
    | psql -h 33.33.33.30 -p 5432 -U nyc_trees

* Removed all existing blockfaces and copied the data into the
survey_blockface table. The original fixtures were segmented by borough
and the id sequence was manipulated so that the leftmost digit of the id
corresponded to a borough. This new fixture is a mix of all boroughs,
but I continued the id pattern for consistency.

    DELETE FROM survey_blockface;
    SELECT setval('survey_blockface_id_seq', 599999);
    INSERT INTO survey_blockface (
      is_available,
      created_at,
      updated_at,
      expert_required,
      source,
      geom
    )
    (
    SELECT
      TRUE,
      now() at time zone 'utc',
      now() at time zone 'utc',
      CASE WHEN expertmapp = 1 THEN TRUE ELSE FALSE END,
      'remaining',
      geom
    FROM "remaining_segments-08-11-15_4326"
    WHERE geom IS NOT NULL
    );

* Used `dumpdata` to create the fixture

    ./manage.py dumpdata survey.Blockface \
    > apps/survey/fixtures/blockface_600000_remaining.json

The fixture will automatically be picked up by the `reload_dev_data`
managment command, but will need to be manually loaded in a production deployment.