Remove all Ansible role related symlinks
Packer complains when it detects symlinks inside of the directory
structure it is responsible for copying to a target EC2 instance. This
step ensures that all of the symlinks are removed before attempting to
run Packer.

All of the symlinks within this directory are being used to help test
roles in isolation.