Fix loading tiles on progress map when logged out
We omitted the dropdown from the progress map when logged out, because
after removing the group view in 83aa2242 there was only a single mode
when not logged in.

This broke loading the tiles on the page, as JS reads the DOM to know
which mode is currently active.  Hiding the element instead of omitting it
fixes the issue while still leaving it not shown for logged out users.

Connects to #1845