Fix map not showing when switching between preview map and normal map
When changing the map element size (e.g if the phone orientation changes),
if the map is not currently visible it will get stuck in an invalid state.

Calling .invalidateSize() when switching between the preview map and the
normal survey map will cause Leaflet to requery the DOM for the map size,
fixing the issue.

Connects to #1867